Check Hibernate is present before loading Hazelcast CacheEntryImpl (#277)
We have to do it, because when Hibernate CacheEntry
is *not* available then loading Hazelcast CacheEntryImpl
throws NoClassDefFoundError.

This makes some application containers (Wildfly) freak
out and log scary warnings. It's just a warning, safe to ignore, but it looks pretty
scary to users.

When Hibernate is NOT available then Class.forName() throws just ClassNotFoundException which we catch anyway. The important difference is that ClassNotFoundException is seen as less severe than NoClassDefFoundError and JBoss/Wildfly won't bother logging it on its own.