- Upgrade hibernate to 6.4.2.Final - The coordinates of hibernate has changed on maven. Use new coordinates - Upgrade version of projects to 5.4.0-SNAPHOT -  AbstractHazelcastCacheRegionFactory needs a new constructor that is : AbstractHazelcastCacheRegionFactory(HazelcastInstance instance)
The reason is Hazelcast spring code is injecting HazelcastInstance with constructor.
The current tests are using hazelcast-hibernate5 dependency which had this constructor. But currently the hazelcast-hibernate5 has been deleted, and it is now hazelcast-hibernate53. But hazelcast-hibernate53 did not bring in the required constructor.

So after this PR is merged, in a new PR, HZ spring tests will be updated to use hazelcast-hibernate53